Added an extra sub Jpanel to each of the eight GUI forms
Added one extra sub Jpanel within the the mainPanel of each form to contain the Jitems and organise them better. Have attempted to start reading in data from the data.csv file (stored in the resources folder) into Jtable for Stockdb form. This is so the file data can be displayed in the table. Might have to change file format from .csv to .txt instead.