Have created two new Jtables on KioskPage.
Have created a products Jtable which lists the products in stock (from stock txt file) for customers to select items to add to their shopping basket. By clicking the "View Products" button a customer can view  the product rows. The customers basket on the right side of the scren is the "basket" Jtable. Here I'm trying to get the functionality working whereby a customer will select a product row from the products Jtable and will then hit the "scan ietm" button which will then transfer the row to the basket Jtable. Still getting this feature to work as of right now.