Made progress on the admin login screen
Have got the success message JLabel to appear in the colour green when entering the corrcet username and password for one of the admin accounts. Have modified the HashMap and learnt about how they work more. Gave two keys which were the two usernames along with their corresponding values which were the relevant passwords for each username. Had made some tests and tampered around with code to get it working. Will now try to clean up code relating to this and will remove any redundant content.