Have added remove product feature in stock db table.
Have added a mouse listener feature to stock JTable so that whenever a product row is selected, the details will appear within the JTextfields for the admin to edit the details of a current product row. Have added remove selected product button "removeBtn" so it removes a product row from the stock JTable.