Tampered around with character spacing for basket list area on KioskPage.
Tampered around with character spacing between Column name values for product details that will be listed in the basket for the customer once they scan a product in with a quantity amount. Have discovered that if a products barcode is more than eight characters in length, the listings will become jagged. Will put an alert jlabel in the StockdbPage that states a barcode being entered for a new product cannot be longer than 8 characters in length. The 8 length rule is the standard rule I have chosen.